+++ shared/versions.ts
@@ -1,4 +1,4 @@
-export const APP_VERSION = "2.3.156";
-export const ANDROID_APP_VERSION = "2.3.44";
-export const ANDROID_APK_RELEASE_VERSION = "2.3.156";
+export const APP_VERSION = "2.3.157";
+export const ANDROID_APP_VERSION = "2.3.45";
+export const ANDROID_APK_RELEASE_VERSION = "2.3.157";
 export const AGENT_VERSION = "2.2.99";